requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
PyJWT==2.9.0
reportlab==4.3.1
sqlparse==0.5.3
tzdata==2025.1
